+++ frontend/src/App.tsx
@@ -1,30 +1,20 @@
-import { BrowserRouter, Routes, Route, Navigate } from 'react-router-dom';
-import HomePage from './pages/HomePage.tsx';
-import LoginPage from './pages/LoginPage.tsx';
-import LogoutPage from './pages/LogoutPage.tsx';
-import UsersPage from './pages/UsersPage.tsx';
-import UsersCreatePage from './pages/UsersCreatePage.tsx';
-import AssetsPage from './pages/AssetsPage.tsx';
-import AssetsCreatePage from './pages/AssetsCreatePage.tsx';
-import AssetPage from './pages/AssetPage.tsx';
-import UserPage from './pages/UserPage.tsx';
-import LoggedRoute from './components/LoggedRoute.tsx';
+import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
+import ROUTES from './config/routes';
+import MainLayout from './layouts/MainLayout';
+import { LoginPage } from './pages/Login';
+import { LogoutPage } from './pages/Logout';
+import { HomePage } from './pages/Home';
 
 export default function () {
     return (
-        <BrowserRouter>
+        <Router>
             <Routes>
-                <Route path='/' element={<HomePage />} />
-                <Route path='/login' element={<LoggedRoute><LoginPage /></LoggedRoute>} />
-                <Route path='/logout' element={<LogoutPage />} />
-                <Route path='/users' element={<UsersPage />} />
-                <Route path='/users/:id' element={<UserPage />} />
-                <Route path='/users/create' element={<UsersCreatePage />} />
-                <Route path='/assets' element={<AssetsPage />} />
-                <Route path='/assets/:id' element={<AssetPage />} />
-                <Route path='/assets/create' element={<AssetsCreatePage />} />
-                <Route path='*' element={<Navigate to='/' replace />} />
+                <Route element={<MainLayout />}>
+                    <Route path={ROUTES.home} element={<HomePage />} />
+                    <Route path={ROUTES.auth.login} element={<LoginPage />} />
+                    <Route path={ROUTES.auth.logout} element={<LogoutPage />} />
+                </Route>
             </Routes>
-        </BrowserRouter>
-    )
-};
+        </Router>
+    );
+}